requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
chromadb==0.4.22
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
